requirements.txt
# MCP framework (ASGI / Streamable HTTP transport)
mcp[cli]>=1.0.0

# ASGI server
uvicorn[standard]>=0.30.0

# HTTP client for Metricool API calls
requests>=2.31.0

# YAML parsing for data-studio-fields.yaml
pyyaml>=6.0.1

# JWT validation (mirrors Java MetricoolJwtDecoder — HMAC-SHA256)
PyJWT>=2.8.0
